feat(ncyBreadcrumbDynamicChain and $Breadcrumb getDynamicStatesChain): Adds a new directive to rendering dynamic and recursive steps.
This directive renders dynamic steps of the breadcrumb.
The property 'ncyBreadcrumb.dynamicStatesChain' in states configuration is a function that returns an array of objects containing the breadcrumb steps.
It is useful when states are parent of itself (recursive) in the breadcrumb, or when the steps come from another service.
For example, in routes with category hierarchy, category({ id : 1 }) > category({ id : 2 }) > product :

```
$stateProvider.state('category', {
  url:'/:id',
  ncyBreadcrumb: {
    dynamicStatesChain: function($injector){
      var myService = $injector.get('myService');

      //The promise resolves with:
      //[{ label: 'Category A', state: 'category({ id : 1 })'},{ label: 'Category B', state: 'category({ id : 2 })'}],
      var promise = myService.getSteps();
      return promise;
    }
  }
}).state('product', {
  url:'/product',
  ncyBreadcrumb: {
    label : 'Product'
  }
});
```

Uses a new function 'getDynamicStatesChain' added on $Breadcrumb service.
Returns a state chain or, if no value is specified for the 'dynamicStatesChain' property, uses the default 'getStatesChain' result.